\name{qmap-package}
\alias{qmap-package}
\alias{qmap}
\docType{package}
\title{
  Quantile-mapping for post-processing climate model output
}
\description{
  Empirical adjustment (bias correction) of variables originating from
  (regional) climate model simulations using quantile mapping. The 
  workhorse functions of this package are \code{\link{fitQmap}} and
  \code{\link{doQmap}} which offer an easy to use interface to different
  quantile mapping methods. 
}
\details{
\tabular{ll}{
Package: \tab qmap\cr
Type: \tab Package\cr
Version: \tab 1.0-0\cr
Date: \tab 2012-03-19\cr
License: \tab GPL >= 2\cr
LazyLoad: \tab yes\cr
}

}
\author{Lukas Gudmundsson}
%% \references{}
\keyword{downscaling, quantile mapping, model output statistics, bias
  correction} 

