% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qMAP.R
\name{qMAP.MH.single_family.1}
\alias{qMAP.MH.single_family.1}
\title{qMAP.MH.single_family.1 - Run qMAP.MH on sncRNA matrix file (file-based input)}
\usage{
qMAP.MH.single_family.1(input_file, ...)
}
\arguments{
\item{input_file}{Path to a tab-delimited read-count matrix. The first two columns should be "Sequence" and "Annotation", and the rest are counts.}

\item{...}{Additional arguments passed to `qMAP.MH.single_family()`: `cl`, `sncrna_family`, `parental_rna_filename`, etc.}
}
\value{
A data frame with results from the qMAP.MH module
}
\description{
This is a variant for 'qMAP.MH.single_family( )' that reads the input matrix from a file.
}
