% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qMAP.R
\name{qMAP.MH.single_family.2}
\alias{qMAP.MH.single_family.2}
\title{qMAP.MH.single_family.2 - Run qMAP.MH on sncRNA data frame (in-memory)}
\usage{
qMAP.MH.single_family.2(input_dataframe, ...)
}
\arguments{
\item{input_dataframe}{A data frame with "Sequence", "Annotation", and read-count columns.}

\item{...}{Additional arguments passed to `qMAP.MH.single_family()`: `cl`, `sncrna_family`, `parental_rna_filename`, etc.}
}
\value{
A data frame with results from the qMAP.MH module
}
\description{
This is a variant for 'qMAP.MH.single_family( )' where the matrix is already in memory as a data frame.
}
