% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qMAP.R
\name{qMAP.MH.single_family}
\alias{qMAP.MH.single_family}
\title{qMAP.MH.single_family - to identify individual sncRNA species that contribute to the observed differential fragmentation of a given parental RNA across groups}
\usage{
qMAP.MH.single_family(
  matrix_file,
  cl,
  sncrna_family,
  parental_rna_filename,
  max.mismatch = getOption("max.mismatch", 1),
  minimum_mean_count = getOption("minimum_mean_count", 5),
  output = NULL
)
}
\arguments{
\item{matrix_file}{A user-provided read-count matrix where the first and second columns are labeled as "Sequence" and "Annotation", respectively. The remaining columns contain read counts from different samples under study.}

\item{cl}{A sample classification vector allowing only two groups: 1 for controls, 2 for cases, and -1 for the samples to be excluded from the analysis.}

\item{sncrna_family}{This parameter specifies a single sncRNA family name to search for in the user-provided parental RNA FASTA file. If a match is found, differential fragmentation analysis will be performed for that family. Use any single family name, such as : “mature-tRNA-Ala-AGC”, “tRNA-Ala-AGC”, “mature-mt_tRNA-His-GTG”, “mt_tRNA-His-GTG”, “5S-rRNA”, “16S-rRNA”}

\item{parental_rna_filename}{A FASTA file containing sequences of parental RNA families description.}

\item{max.mismatch}{This parameter specifies the maximum number of allowed mismatches (default is 1) when mapping sequence reads onto the FASTA sequence of the parental RNA.}

\item{minimum_mean_count}{This parameter sets the minimum allowed mean read count (default: 5) across samples for the individual sncRNA species. Only sncRNA species (i.e., rows in the read-count matrix)  with mean read counts greater than or equal to minimum_mean_count will be included in the analysis.}

\item{output}{This parameter specifies the name of the output file, By default, if no name is provided,  the output file will be saved in the working directory with a name like “qMAP_MH_output_X.txt", where X represents query sncrna_family parameter used in the run. If '"temp"', the function creates a temporary output file (used for examples/testing only).}
}
\value{
It will return, for each queried parental RNA (sncrna_family), an output data-frame with sequence (the sncRNA species sequence), or (the odds ratio), p (the P-value inferred from the Mantel-Haenszel statistic), and start (the start position on the parental RNA where the sncRNA species is mapped).
}
\description{
qMAP.MH.single_family - to identify individual sncRNA species that contribute to the observed differential fragmentation of a given parental RNA across groups
}
\note{
In example runs, `output = "temp"` is used to avoid writing files during checks. In real analyses, you may provide a file path or let the function auto-name the output (see manual).
}
\examples{
# Note:  After installation, one can find the example files in "../qMAPModules/extdata/"
# To run on a specific input, provide the full path as: qMAPModules(a = "/path/to/your_matrix.txt")
sample_matrix_1 <- system.file("extdata", "sample_matrix1.txt", package = "qMAPModules")
sample_matrix_2 <- system.file("extdata", "sample_matrix2.txt", package = "qMAPModules")
sample_matrix_3 <- system.file("extdata", "sample_matrix3.txt", package = "qMAPModules")
mouse_gtsrna_list <- system.file("extdata", "mmu_gtsrna.txt", package = "qMAPModules")
mouse_rRNAs <- system.file("extdata", "mouse_rRNAs.fa", package = "qMAPModules")
mouse_tRNAs <- system.file("extdata", "mm10-tRNAs_CCA.fa", package = "qMAPModules")
# Run the identification of sncRNA species contributing to the differential fragmentation of 5S-rRNA
# family using corresponding FASTA sequence in the user provided parental rna file (mouse_rRNAs.fa)
# by mapping sequence reads in the query read-count matrix file (sample_matrix1.txt) comprising read
# counts for 5 control (cl=1) and 4 case (cl=2) samples, with allowed maximum mismatch 1 (default)
# and considering only the reads with the row mean read-counts across the samples > 5 (default).
qMAP.MH.single_family.1(input_file = sample_matrix_1, cl=c(1,1,1,1,1,2,2,2,2),
                    sncrna_family="5S-rRNA", parental_rna_filename= mouse_rRNAs, output = "temp")

# Run the identification of sncRNA species contributing to the differential fragmentation of
# tRNA-Gly-CCC family using corresponding FASTA sequence in the user provided parental rna file
# (mm10-tRNAs_CCA.fa) by mapping sequence reads in the query read-count matrix file
# (sample_matrix1.txt) comprising read counts for 3 control (first three sample columns in the
# matrix) (cl = 1) and 3 case (last three sample columns in the matrix) (cl = 2) samples, with
# allowed maximum mismatch 1 (default) and considering only the reads with the row mean read-counts
# across the samples > 10. Here, 3 samples are excluded (cl = -1).
qMAP.MH.single_family.1(sample_matrix_2, cl=c(1,1,1,-1,-1,-1,2,2,2), sncrna_family="tRNA-Gly-CCC",
                  minimum_mean_count  = 10,  parental_rna_filename=mouse_tRNAs, output = "temp")

# Run the identification of sncRNA species contributing to the differential fragmentation of
# tRNA-Val-AAC family using corresponding FASTA sequence in the user provided parental rna file
# (mm10-tRNAs_CCA.fa) by mapping sequence reads in the query read-count matrix file
# (sample_matrix3.txt) comprising read counts for 4 Contioin1 (cl = 1), 4 Condition2 (cl =2)
# samples, and rest 8 for other conditions, with allowed maximum mismatch 0 and considering only
# the reads with the row mean read-counts across the samples > 10. The result output will be saved
# to the specified file.
qMAP.MH.single_family.1(sample_matrix_3, cl=c(1,1,1,1,2,2,2,2,-1,-1,-1,-1,-1,-1,-1,-1),
                  sncrna_family="tRNA-Val-AAC", minimum_mean_count  = 10, max.mismatch = 0,
                  parental_rna_filename = mouse_tRNAs, output = "temp")

# Running with query read-count matrix as a data frame as input.
# First read and save the read-count matrix file as a dataframe:
a <- read.delim(sample_matrix_1, header = TRUE)
# Then call qMAP_mh function to run the identification of sncRNA species contributing to the
# differential fragmentation of tRNA-His-GTG family using corresponding FASTA sequence in the user
# provided parental rna file (mm10-tRNAs_CCA.fa) by mapping sequence reads in the query read-count
# matrix file (sample_matrix1.txt) comprising read counts for 5 control (cl = 1) and 4 case (cl =2)
# samples, with allowed maximum mismatch 1 (default) and considering only the reads with the row
# mean read-counts across the samples > 5 (default).
qMAP.MH.single_family.2(a, cl=c(1,1,1,1,1,2,2,2,2), sncrna_family="tRNA-His-GTG",
		 	parental_rna_filename= mouse_tRNAs, output ="temp")

# Run the identification of sncRNA species contributing to the differential fragmentation of each of
# the RNA families (listed in a file; here it is mmu_gtsrna.txt) using corresponding FASTA sequences
# in the user provided parental rna file (mm10-tRNAs_CCA.fa) by mapping sequence reads in the query
# read-count matrix file (sample_matrix2.txt) comprising read counts for 3 control (first three
# sample columns in the matrix) (cl = 1) and 3 case (next three sample columns in the matrix) (cl=2)
# samples. Here, last 3 samples excluded (cl = -1). It will run with allowed maximum mismatch 1
# (default) and considering only the reads with the row mean read-counts across the samples > 5
# (default). The result output will be saved to the specified file, qmap.mh_multifamily_results.txt.
# To run qMAP_mh for multiple families, it is suggested to run with second variant
# (qMAP.MH.single_family.2) in a loop as follow:
# First read and save the read-count matrix file as a dataframe:
a = read.delim(sample_matrix_2)
cl = c(1,1,1,2,2,2,-1,-1,-1)
result = c()
# Read sncRNA families from list in the file:
sncrna_family_list = read.table(mouse_gtsrna_list)[,1]
# Then for each family in the list, call qMAP_mh function
for (i in 1:length(sncrna_family_list)) {
    result = rbind(result, qMAP.MH.single_family.2(a, cl=cl, sncrna_family=sncrna_family_list[i],
    parental_rna_filename = mouse_tRNAs, output = "temp"))
  }
# Creating a safe temporary output file exclusively for run examples (see manual)
out_file <- tempfile("qmap.mh_multifamily_results", fileext = ".txt")
write.table(result, file = out_file, quote=FALSE, row.names=FALSE, sep="\t")

}
