% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qMAP.R
\name{qMAP}
\alias{qMAP}
\title{qMAP - quantitative mapping for RNA differential fragmentation}
\arguments{
\item{matrix_file}{A user-provided read-count matrix where the first and second columns are labeled as "Sequence" and "Annotation", respectively. The remaining columns contain read counts from different samples under study.}

\item{cl}{A sample classification vector allowing only two groups: 1 for controls, 2 for cases, and -1 for the samples to be excluded from the analysis.}

\item{parental_rna_filename}{A FASTA file containing sequences of parental RNA families description.}

\item{sncrna_family}{This parameter specifies a single sncRNA family name to search for in the user-provided parental RNA FASTA file. If a match is found, differential fragmentation analysis will be performed for that family. Use any single family name, such as : “mature-tRNA-Ala-AGC”, “tRNA-Ala-AGC”, “mature-mt_tRNA-His-GTG”, “mt_tRNA-His-GTG”, “5S-rRNA”, “16S-rRNA”}

\item{max.mismatch}{This parameter specifies the maximum number of allowed mismatches (default is 1) when mapping sequence reads onto the FASTA sequence of the parental RNA.}
}
\value{
with qMAP.single_family, it will return, for each queried parental RNA (sncrna_family), an output data-frame with  mean_diff (mean difference in coverage between two groups), p (the P-value indicating statistical significance ), and species_num (number of contributing sncRNA species mapped to the parental RNA).

with qMAP.MH.single_family, it will return, for each queried parental RNA (sncrna_family), an output data-frame with sequence (the sncRNA species sequence), or (the odds ratio), p (the P-value inferred from the Mantel-Haenszel statistic), and start (the start position on the parental RNA where the sncRNA species is mapped).
}
\description{
qMAP - quantitative mapping for RNA differential fragmentation
}
\details{
For more information, see the user manual:
\code{browseURL(system.file("docs/qMAP_v1.0.0_User_Manual.pdf", package = "qMAP"))}

After installing the package qMAP, call the library as: library(qMAP).

There are two modules or main functions in this package : qMAP and qMAP_mh. qMAP has two model/methods: Model1 and Model2

Ask help for the description and help menus of each of these functions :
\cr ?qMAP.single_family
\cr ?qMAP.MH.single_family

If want to run examples, it is necessary to set working directory to the base folder of the installed package qMAP. Such as : setwd("/home/..../R/.../qMAP/").
}
