% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qMAP.R
\name{qMAP.single_family}
\alias{qMAP.single_family}
\title{qMAP.single_family - identification of differentially fragmented parental RNAs between groups through quantitative mapping of sncRNAs}
\usage{
qMAP.single_family(
  matrix_file,
  cl,
  sncrna_family,
  parental_rna_filename,
  max.mismatch = getOption("max.mismatch", 1),
  min_count = getOption("min_count", 10),
  shuffling_round = NULL,
  method = 1,
  output = NULL
)
}
\arguments{
\item{matrix_file}{A user-provided read-count matrix where the first and second columns are labeled as "Sequence" and "Annotation", respectively. The remaining columns contain read counts from different samples under study.}

\item{cl}{A sample classification vector allowing only two groups: 1 for controls, 2 for cases, and -1 for the samples to be excluded from the analysis.}

\item{sncrna_family}{This parameter specifies a single sncRNA family name to search for in the user-provided parental RNA FASTA file. If a match is found, differential fragmentation analysis will be performed for that family. Use any single family name, such as : “mature-tRNA-Ala-AGC”, “tRNA-Ala-AGC”, “mature-mt_tRNA-His-GTG”, “mt_tRNA-His-GTG”, “5S-rRNA”, “16S-rRNA”}

\item{parental_rna_filename}{A FASTA file containing sequences of parental RNA families description.}

\item{max.mismatch}{This parameter specifies the maximum number of allowed mismatches (default is 1) when mapping sequence reads onto the FASTA sequence of the parental RNA.}

\item{min_count}{This parameter sets the minimum allowed mean read-count across samples for the sncRNA family to be considered in the analysis. The default value is 10, meaning only families with mean read-counts greater than or equal to 10 across samples will be included.}

\item{shuffling_round}{This parameter specifies the minimum number of sample shuffling rounds performed during the statistical model to identify differential fragmentation. The default value is 100 for Model 1 (method =1),  and 1000 for Model 2 (method =2)).}

\item{method}{Specify either “1” or “2” to select which model to run for identifying differentially fragmented parental RNAs. Use “2” to perform analysis with Model 2 (method = 2) or use “1” (default) to run Model 1 (method = 1).}

\item{output}{This parameter specifies the name of the output file, By default, if no name is provided,  the output file will be saved in the working directory with a name like “qMAP_****_output.txt", where * represents various  parameters used in the run. If '"temp"', the function creates a temporary output file (used for examples/testing only).}
}
\value{
with qMAP, it will return, for each queried parental RNA (sncrna_family), an output data-frame with  mean_diff (mean difference in coverage between two groups), p (the P-value indicating statistical significance ), and species_num (number of contributing sncRNA species mapped to the parental RNA).
}
\description{
qMAP.single_family - identification of differentially fragmented parental RNAs between groups through quantitative mapping of sncRNAs
}
\note{
In example runs, `output = "temp"` is used to avoid writing files during checks. In real analyses, you may provide a file path or let the function auto-name the output (see manual).
}
\examples{
# Note:  After installation, one can find the example files in "../qMAPModules/extdata/"
# To run on a specific input, provide the full path as: qMAPModules(a = "/path/to/your_matrix.txt")
sample_matrix_1 <- system.file("extdata", "sample_matrix1.txt", package = "qMAPModules")
sample_matrix_2 <- system.file("extdata", "sample_matrix2.txt", package = "qMAPModules")
sample_matrix_3 <- system.file("extdata", "sample_matrix3.txt", package = "qMAPModules")
mouse_gtsrna_list <- system.file("extdata", "mmu_gtsrna.txt", package = "qMAPModules")
mouse_rRNAs <- system.file("extdata", "mouse_rRNAs.fa", package = "qMAPModules")
mouse_tRNAs <- system.file("extdata", "mm10-tRNAs_CCA.fa", package = "qMAPModules")
# Run differential fragmentation analysis for 5S-rRNA family using  corresponding FASTA sequence
# in the user provided parental rna file (mouse_rRNAs.fa) by mapping sequencing reads in the query
# read-count matrix file (sample_matrix1.txt) comprising read counts for 5 control (cl = 1) and
# 4 case (cl =2) samples. It will run this for method 1 (default)  with allowed maximum mismatch 1
# (default) and with 100 shuffling rounds (default, for method 1), only if the mean total read-
# counts across the samples for the mapped reads (against query sncRNA family) > 10 (default).
qMAP.single_family.1(input_file = sample_matrix_1, cl=c(1,1,1,1,1,2,2,2,2),
                    sncrna_family="5S-rRNA", parental_rna_filename= mouse_rRNAs, output = "temp")

# Run differential fragmentation analysis in a similar manner as for example 1 above.
qMAP.single_family.1(sample_matrix_1, cl=c(1,1,1,1,1,2,2,2,2), sncrna_family="5S-rRNA",
		parental_rna_filename=mouse_rRNAs, method = "1", max.mismatch =1,
		shuffling_round = 100, min_count = 10, output = "temp")

# Run differential fragmentation analysis for 16S-rRNA family using  corresponding FASTA sequence
# in the user provided parental rna file (mouse_rRNAs.fa) by mapping sequence reads in the query
# read-count matrix file (sample_matrix2.txt) comprising read counts for 3 control (first three
# sample columns in the matrix) (cl = 1) and 3 case (last three sample columns in the matrix)
# (cl = 2) samples. Here, 3 samples excluded (cl = -1). It will run this for method 1 (default) with
# allowed maximum mismatch 1 (default) and with 100 shuffling rounds (default, for method 1), only
# if the mean total read-counts across the samples for the mapped reads (against query sncRNA
# family) > 10 (default). The result output will be saved to the specified file.
qMAP.single_family.1(sample_matrix_2, cl=c(1,1,1,-1,-1,-1,2,2,2), sncrna_family="16S-rRNA",
			parental_rna_filename = mouse_rRNAs,  output = "temp")

# Run differential fragmentation analysis for tRNA-Gly-CCC family by applying method 2 using
# corresponding FASTA sequence in the user provided parental rna file (mm10-tRNAs_CCA.fa) by mapping
# sequence reads in the query read-count matrix file (sample_matrix1.txt) comprising read counts for
# 5 control (cl = 1) and 4 case (cl = 2) samples. It will run this with allowed maximum mismatch 1
# (default) and with 1000 shuffling rounds (default, for method 2), only if the mean total read-
# counts across the samples for the mapped reads (against query sncRNA family) > 10 (default).
qMAP.single_family.1(sample_matrix_1, cl=c(1,1,1,1,1,2,2,2,2), sncrna_family="tRNA-Gly-CCC",
			parental_rna_filename = mouse_tRNAs, method = "2", output = "temp")

# Run differential fragmentation analysis for tRNA-Gly-CCC family by applying method 2 using
# corresponding FASTA sequence in the user provided parental rna file (mm10-tRNAs_CCA.fa) by mapping
# sequence reads in the query read-count matrix file (sample_matrix1.txt) comprising read counts for
# 5 control (cl = 1) and 4 case (cl = 2) samples. It will run this with allowed maximum mismatch 1
# (default) and with 1000 shuffling rounds (default, for method 2), only if the mean total read-
# counts across the samples for the mapped reads (against query sncRNA family) > 2.
qMAP.single_family.1(sample_matrix_1, cl=c(1,1,1,1,1,2,2,2,2), sncrna_family="tRNA-Gly-CCC",
		parental_rna_filename = mouse_tRNAs, min_count = 2, method = "2", output = "temp")

# Run differential fragmentation analysis for tRNA-Val-AAC family by applying method 2 using
# corresponding FASTA sequence in the user provided parental rna file (mm10-tRNAs_CCA.fa) by
# mapping sequence reads (with mismatch zero) in the query read-count matrix file
# (sample_matrix3.txt) comprising read counts for 4 Contioin1 (cl = 1), 4 Condition2 (cl = 2)
# samples, and rest 8 for other conditions. It will run this with 1000 shuffling rounds (default,
# for method 2), only if the mean total read-counts across the samples for the mapped reads (against
# query sncRNA family) > 10 (default).
qMAP.single_family.1(sample_matrix_3, cl=c(1,1,1,1,2,2,2,2,-1,-1,-1,-1,-1,-1,-1,-1),
			sncrna_family="tRNA-Val-AAC", parental_rna_filename = mouse_tRNAs,
			max.mismatch = 0, method = "2", output = "temp")

# Running with query read-count matrix as a data frame as input
# First read and save the read-count matrix file as a dataframe:
a <- read.delim(sample_matrix_1, header = TRUE)
# Then call qMAP function to run differential fragmentation analysis for tRNA-His-GTG family by
# applying method 2 using corresponding FASTA sequence in the user provided parental rna file
# (mm10-tRNAs_CCA.fa) by mapping sequence reads (with mismatch 1) in the query read-count matrix
# file (a, sample_matrix1.txt) comprising read counts for 5 control (cl = 1) and 4 case (cl =2)
# samples. It will run this with 1000 shuffling rounds (default, for method 2), only if the mean
# total read-counts across the samples for the mapped reads (against query sncRNA family) > 10
# (default).
qMAP.single_family.2(a, cl=c(1,1,1,1,1,2,2,2,2), sncrna_family="tRNA-His-GTG",
		parental_rna_filename= mouse_tRNAs, max.mismatch = 1, method = "2", output = "temp")

# Run differential fragmentation analysis for each of the RNA families (listed in a file; here it
# is mmu_gtsrna.txt) using  corresponding FASTA sequences in the user provided parental rna file
# (mm10-tRNAs_CCA.fa) by mapping sequence reads in the query read-count matrix file
# (sample_matrix2.txt) comprising read counts for 3 control (first three sample columns in the
# matrix) (cl = 1) and 3 case (next three sample columns in the matrix) (cl = 2) samples. Here,
# last 3 samples excluded (cl = -1). It will run this for method 1 with allowed maximum mismatch 1
# (default) and with 100 shuffling rounds, only if the mean total read-counts across the samples
# for the mapped reads (against query sncRNA family) > 10 (default). The result output will be
# saved to the specified file, qmap_multifamily_results.txt.
# To run qMAP for multiple families, it is suggested to run with the second variant
# (qMAP.single_family.2) in a for loop as follows:
# First read and save the read-count matrix file as a dataframe
a = read.delim(sample_matrix_2)
cl = c(1,1,1,2,2,2,-1,-1,-1)
result = c()
# Read sncRNA families from list in the file
sncrna_family_list = read.table(mouse_gtsrna_list)[,1]
# Then for each family in the list, call qMAP function
for (i in 1:length(sncrna_family_list)) {
	result = rbind(result, qMAP.single_family.2(a, cl=cl, sncrna_family=sncrna_family_list[i],
	parental_rna_filename = mouse_tRNAs, method=1, shuffling_round=100, output="temp"))
	}
# Creating a safe temporary output file exclusively for run examples (see manual)
out_file <- tempfile("qmap_multifamily_results", fileext = ".txt")
write.table(result, file = out_file, quote=FALSE, row.names=FALSE, sep="\t")

}
