\name{qmap-package}
\alias{qmap-package}
\alias{qmap}
\docType{package}
\title{
  Statistical Transformations for Post-Processing Climate Model Output
}
\description{
  Empirical adjustment (bias correction) of variables originating from
  (regional) climate model simulations using quantile mapping. The 
  workhorse functions of this package are \code{\link{fitQmap}} and
  \code{\link{doQmap}} which offer an easy to use interface to different
  statistical transformations, also referred to as quantile mapping
  methods.  
}
\details{
\tabular{ll}{
Package: \tab qmap\cr
Type: \tab Package\cr
Version: \tab 1.0-4\cr
Date: \tab 2016-05-03\cr
License: \tab GPL >= 2\cr
LazyLoad: \tab yes\cr
}
}
\author{Lukas Gudmundsson}
\references{
  Gudmundsson, L.; Bremnes, J. B.; Haugen, J. E. & Engen-Skaugen,
  T. Technical Note: Downscaling RCM precipitation to the station scale
  using statistical transformations - a comparison of methods. Hydrology
  and Earth System Sciences, 2012, 16, 3383-3390,
  <doi:10.5194/hess-16-3383-2012>.  
}
\keyword{distribution}
\concept{quantile mapping}
\concept{quantile matching}
\concept{cumulative distribution function (cdf) matching}
\concept{quantile - quantile transformation}
\concept{histogram equalisation}
\concept{histogram matching}
\concept{probability mapping} 
\concept{distribution mapping}
\concept{Probability integral transform}
\concept{model output statistics}
\concept{bias correction}
\concept{downscaling}


