\name{qmap-package}
\alias{qmap-package}
\alias{qmap}
\docType{package}
\title{
  Statistical transformations for post-processing climate model output
}
\description{
  Empirical adjustment (bias correction) of variables originating from
  (regional) climate model simulations using quantile mapping. The 
  workhorse functions of this package are \code{\link{fitQmap}} and
  \code{\link{doQmap}} which offer an easy to use interface to different
  statistical transformations, also referred to as quantile mapping
  methods.  
}
\details{
\tabular{ll}{
Package: \tab qmap\cr
Type: \tab Package\cr
Version: \tab 1.0-3\cr
Date: \tab 2014-06-10\cr
License: \tab GPL >= 2\cr
LazyLoad: \tab yes\cr
}

}
\author{Lukas Gudmundsson}
\references{
  Gudmundsson, L.; Bremnes, J. B.; Haugen, J. E. & Engen-Skaugen,
  T. Technical Note: Downscaling RCM precipitation to the station scale
  using statistical transformations - a comparison of methods. Hydrology
  and Earth System Sciences, 2012, 16, 3383-3390,
  doi:10.5194/hess-16-3383-2012.  
}
\keyword{quantile mapping, quantile matching, cumulative distribution
  function (cdf) matching, quantile - quantile transformation, histogram
  equalisation, histogram matching, probability mapping, distribution
  mapping, model output statistics, bias correction, downscaling}  

